Factored out StoreState from Store
